requirements.txt
absl-py==2.2.2
affine==2.4.0
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
botocore==1.37.3
bounded-pool-executor==0.0.3
Brotli==1.1.0
bs4==0.0.2
Cartopy==0.24.1
certifi==2025.4.26
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-yasg==1.21.10
earthaccess==0.14.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.6.0
geopandas==1.0.1
gevent==24.11.1
geventhttpclient==2.3.3
git-filter-repo==2.47.0
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h5netcdf==1.6.1
h5py==3.13.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
locust==2.37.4
locust-cloud==1.21.7
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.4.3
multimethod==2.0
namex==0.0.9
netCDF4==1.7.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
pqdm==0.2.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
requests==2.32.3
rich==14.0.0
rioxarray==0.19.0
s3fs==2025.3.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.4.0
shapely==2.1.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tinynetrc==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xarray==2025.1.2
xlrd==2.0.1
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
